# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PetalPath Backend API
  description: >
    Adaptive learning platform API for children's educational journey.
    Authentication via JWT bearer token. All protected endpoints require
    the `Authorization: Bearer <token>` header.
  version: 2.0.0
  contact:
    name: PetalPath Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          description: Response payload
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
...
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                enum: [true]
              message:
                type: string
...
    BadRequest:
      description: Bad request / validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized (missing or invalid token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /placement/questionnaire:
    get:
      tags: [Placement]
      summary: "Get placement questionnaire"
      operationId: placementGetQuestionnaire
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Questionnaire"
          content:
            application/json:
              schema:
                allOf:
                  - 
                    $ref: '#/components/schemas/ApiResponse'
                  - 
                    type: object
                    properties:
                      data:
                        type: object
                        description: "Placement questionnaire data"
        '401':
          $ref: '#/components/responses/Unauthorized'
  /placement/start:
    post:
      tags: [Placement]
      summary: "Start placement test"
      operationId: placementStart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childId]
              properties:
                childId:
                  type: "string"
                  format: "uuid"
      responses:
        '201':
          description: "Placement started"
          content:
            application/json:
              schema:
                allOf:
                  - 
                    $ref: '#/components/schemas/ApiResponse'
                  - 
                    type: object
                    properties:
                      data:
                        type: object
                        description: "Placement session created"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /placement/start-from-beginning:
    post:
      tags: [Placement]
      summary: "Restart placement from beginning"
      operationId: placementStartFromBeginning
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childId]
              properties:
                childId:
                  type: "string"
                  format: "uuid"
      responses:
        '201':
          description: "Placement restarted"
          content:
            application/json:
              schema:
                allOf:
                  - 
                    $ref: '#/components/schemas/ApiResponse'
                  - 
                    type: object
                    properties:
                      data:
                        type: object
                        description: "Placement session restarted"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /placement/children/{childId}/answer:
    post:
      tags: [Placement]
      summary: "Submit placement answer"
      operationId: placementSubmitAnswer
      security:
        - bearerAuth: []
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionId, answer]
              properties:
                questionId:
                  type: "string"
                answer:
                timeSpent:
                  type: "integer"
      responses:
        '200':
          description: "Answer submitted"
          content:
            application/json:
              schema:
                allOf:
                  - 
                    $ref: '#/components/schemas/ApiResponse'
                  - 
                    type: object
                    properties:
                      data:
                        type: object
                        description: "Answer result"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /placement/children/{childId}/complete:
    post:
      tags: [Placement]
      summary: "Complete placement test"
      operationId: placementComplete
      security:
        - bearerAuth: []
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /placement/children/{childId}/result/{attemptId}:
    get:
      tags: [Placement]
      summary: "Get placement result"
      operationId: placementGetResult
      security:
        - bearerAuth: []
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: "Placement result"
          content:
            application/json:
              schema:
                allOf:
                  - 
                    $ref: '#/components/schemas/ApiResponse'
                  - 
                    type: object
                    properties:
                      data:
                        type: object
                        description: "Placement result data"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /placement/children/{childId}/restart:
    post:
      tags: [Placement]
      summary: "Restart placement test"
      operationId: placementRestart
      security:
        - bearerAuth: []
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...